Update README.md
This guidance is intended for those who are setting up new firewalls for small enterprises or home pro environments. It assumes the use of low-cost commercial firewalls that are common in that environment such as pfSense, OPNsense, Sophos, etc.  The guidance below assumes pfSense as the default. 

The guidance is also based on a simple, somewhat generic environment. For business in particular you will need to customize it based on the particular requirements of your business for segmentation, permanent VPN tunnels, and web filtering, none of which are included below.